requirements.txt
Django==1.10.2
django-appconf==1.0.1
django-image-cropping==1.0.3
django-resized==0.3.5
easy-thumbnails==2.3
Pillow==3.4.2
six==1.10.0
sorl-thumbnail==12.4a1
